requirements.txt
aiohttp
discord.py
google-generativeai
python-dotenv
youtube-transcript-api
PyMuPDF
requests
beautifulsoup4